requirements.txt
matplotlib
numpy<2
opencv_python
pycocotools
torch
torchvision
tqdm

onnxsim
sam4onnx
ssc4onnx
onnx2json
json2onnx
jq
